refactor: index all AddEditForm components
